Extract phantom type definitions from resource.rs to resource/types.rs
First step in breaking down the 1,415-line resource.rs file. Extracted
all phantom type markers (Acquired, Released, Initializing, Failed) and
their associated traits (IsAcquired, IsReleasable, IsRecoverable) into
a dedicated types module.

This separation improves code organization and makes the type-state
pattern more explicit. The sealed trait pattern prevents external code
from implementing these marker traits, maintaining API safety.

All tests pass with the new module structure.
