+++ PLANNING.md
@@ -69,7 +69,7 @@ EventCore has successfully completed all initially planned phases (1-20), includ
 #### 3. Refactor resource.rs (1,415 lines) - **HIGH**
 **Problem**: Single file handling all resource lifecycle patterns
 **Tasks**:
-- [ ] Extract phantom type definitions to `resource/types.rs`
+- [x] Extract phantom type definitions to `resource/types.rs`
 - [ ] Move concrete implementations to `resource/implementations.rs`
 - [ ] Create `resource/lifecycle.rs` for acquisition/release patterns
 - [ ] Create `resource/pool.rs` for resource pooling
